// src/app/interfaces/contact-info.ts
export interface ContactInfo {
  _id?: string;
  primaryPhoneNo?: string;
  secondaryPhoneNo?: string;
  addressLineOne?: string;
  addressLineTwo?: string;
  addressLines?: string[];
  email?: string;
  facebookLink?: string;
  youtubeLink?: string;
  instagramLink?: string;
  twitterLink?: string;
  shopOpenTime?: string;
  notification?: string;
}